utils/requirements.txt
numba
opt-einsum
numpy
absl-py
termcolor
matplotlib
ml-collections
cython<3
wandb
imageio
moviepy
opensimplex
pygame
libtmux
threadpoolctl==3.1.0
plotly
tabulate
tensorflow-cpu
tensorflow-probability==0.22.0
tensorflow-datasets
diffusers
fabric
opt-einsum
scipy==1.12.0
einops
jaxtyping

flax==0.7.4
optax==0.1.7
orbax==0.1.9
distrax==0.1.4
chex==0.1.82
